Moving model tests from API tests file to more proper location.
